il frutto di qualche razionalizzazione

+++ vbox/viewpoint/setup-ssh-agent.sh
@@ -1,6 +1,10 @@
 # add the key used by GO pipelines
 # to use fleetctl remotly
 
-rm /tmp/GO_SSH_ANGENT.sck
+if [ -f /tmp/GO_SSH_ANGENT.sck ]
+then
+  rm /tmp/GO_SSH_ANGENT.sck
+fi
+
 eval `ssh-agent -s -a /tmp/GO_SSH_ANGENT.sck`
 ssh-add /anarchia/certs/id_rsa_anarchia_go
